pyproject.toml
[project]
name = "homalos-datacenter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.1",
    "fastapi>=0.120.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "psutil>=7.1.2",
    "pydantic[email]>=2.12.3",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.38.0",
]
